+++ R/utils.R
@@ -130,7 +130,7 @@
 #reshuffling df
 #' @keywords internal
 .groupList <- function(df, group.by) {
-    df <- bind_rows(df)
+    df <- do.call(rbind, df)
     df <- split(df, df[,group.by])
     return(df)
 }