+++ NEWS.md
@@ -21,6 +21,7 @@ nt sequences. No longer calculates all possible motifs, but only motifs present.
 * ```clonalCompare()``` issue with plotting a 0 row data frame now errors with message
 * ```clonalScatter()``` group.by/axes call now works for non-single-cell objects
 * Fixed issue with NULL and "none" group.by in ```combineExpression()```
+* Allowing multi groupings via x.axis and group.by in ```clonalDiversity()```
 
 # scRepertoire VERSION 2.3.4